Code Review updates
