+++ nhd/Node.py
@@ -95,6 +95,9 @@ The Node class holds properties about a node's resources, as well as which resou
 Current resource types in a node are CPUs, GPU, and NICs.
 """
 class Node:
+
+    NHD_MAINT_LABEL = 'sigproc.viasat.io/maintenance'
+
     def __init__(self, name, active = True):
         self.logger = NHDCommon.GetLogger(__name__)
 
@@ -115,6 +118,16 @@ class Node:
         self.gwip : str = '0.0.0.0/32'
         self.mem: NodeMemory = NodeMemory()
         self.reserved_cores = [] # Reserved CPU cores
+
+    def GetMaintenance(labels):
+        maintenance = False
+        if (Node.NHD_MAINT_LABEL in labels):
+            value = labels[Node.NHD_MAINT_LABEL].lower()
+            if (value != 'not_scheduled'):
+                maintenance = True
+
+        return maintenance
+
 
     def ResetResources(self):
         """ Resets all resources back to initial values """
@@ -298,12 +311,7 @@ class Node:
         return True    
 
     def InitMaintenance(self, labels):
-        self.maintenance = False
-        if (NHDCommon.NHD_MAINT_LABEL in labels):
-            value = labels[NHDCommon.NHD_MAINT_LABEL].lower()
-            if (value != 'no'):
-                self.maintenance = True
-    
+        self.maintenance = Node.GetMaintenance(labels)
         return True    
 
     def InitCores(self, labels):
