+++ nhd/TriadController.py
@@ -9,6 +9,7 @@ from nhd.NHDWatchQueue import qinst
 from nhd.NHDWatchQueue import NHDWatchTypes
 from nhd.NHDCommon import NHDCommon
 from nhd.NHDScheduler import NHD_SCHED_NAME
+from nhd.Node import Node
 import kopf
 import yaml
 import os
@@ -33,15 +34,6 @@ def TriadSetCreate(spec, meta, **_):
 def delete_fn(meta, **_):
     logger = NHDCommon.GetLogger(__name__)
     logger.info('Received delete request for TriadSet')
-
-def GetMaintenance(labels):
-    maintenance = False
-    if (NHDCommon.NHD_MAINT_LABEL in labels):
-        value = labels[NHDCommon.NHD_MAINT_LABEL].lower()
-        if (value != 'no'):
-            maintenance = True
-
-    return maintenance
 
 
 # Dete node changes to determine if a node is cordoned and/or the NHD group label is changing
@@ -71,8 +63,8 @@ def TriadNodeUpdate(spec, old, new, meta, **_):
        k8sq.put({"type": NHDWatchTypes.NHD_WATCH_TYPE_GROUP_UPDATE, "node": meta["name"], "groups" : "default"})
 
     # Detect change in node maintenance state
-    oldMaintenance = GetMaintenance(old['metadata']['labels'])
-    newMaintenance = GetMaintenance(new['metadata']['labels'])
+    oldMaintenance = Node.GetMaintenance(old['metadata']['labels'])
+    newMaintenance = Node.GetMaintenance(new['metadata']['labels'])
     if (not oldMaintenance and newMaintenance):
         logger.info(f'Starting Maintenance for node {meta["name"]}')
         k8sq.put({"type": NHDWatchTypes.NHD_WATCH_TYPE_NODE_MAINT_START, "node": meta["name"]})
